+++ CHANGES.md
@@ -1,3 +1,5 @@
+# Unreleased
+- [IMPROVED] Add HTTP status code to `Response` objects.
 # 2.0.0 (2015-11-12)
 - [NEW] `DesignDocument.MapReduce` now has a setter for the `dbcopy` field.
 - [NEW] Requests for the `_all_docs` endpoint are made via `Database#getAllDocsRequestBuilder()`